pyproject.toml
[tool.poetry]
name = "real-state-prices"
version = "0.1.0"
description = "Application for real state price prediction in São Pauo(SP)-BR"
license = "MIT"
authors = ["Lucas NG <garcia@outlook.com"]

[tool.poetry.dependencies]
python = "^3.8"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
numpy = "^1.24"
notebook = "^6.4.10"
pre-commit = "^2.19.0"
dagshub = "^0.1.8"
bentoml = "^1.0.7"
dvc = "^2.31.0"
scrapy = "^2.7.1"
scrapy-playwright = "^0.0.22"
duckdb = "^0.6.0"
pyarrow = "^10.0.1"
jupyterlab = "^3.5.0"
duckdb-engine = "^0.6.5"
polars = "^0.15.1"
rise = "^5.7.1"
jupyter-nbextensions-configurator = "^0.6.1"
seaborn = "^0.12.1"
matplotlib = "^3.6.2"
prefect = "^2.7.8"
prefect-jupyter = "^0.2.0"
joblib = "^1.2.0"
interrogate = "^1.5.0"
scikit-learn = "^1.2.0"
xgboost = "^1.7.3"
mlflow = "^2.1.1"
pydantic = "^1.10.4"
lightgbm = "^3.3.5"
sweetviz = "^2.1.4"
scikit-optimize = {git = "https://github.com/scikit-optimize/scikit-optimize.git"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

#[tool.black]
#line-length = 79
#include = '\.pyi?$'
#exclude = '''
#/(
#	\.git
#| \.hg
#| \.mypy_cache
#| \.tox
#| \.venv
#| _build
#| buck-out
#| build   
#)/ 
#'''

[tools.ruff]
extend-select = [
	"B",
	"BLE",
	"C4",
	"ERA",
	"I",
	"ICN",
	"ISC",
	"N",
	"NPY",
	"PGH",
	"PIE",
	"Q",
	"RET",
	"RSE",
	"RUF",
	"S",
	"SIM",
	"T20",
	"TCH",
	"TUD",
	"YTT",
	"ANN201",
	"ANN202",
	"ANN204",
	"ANN205",
	"ANN206",
	"E501"
]
ignore = ["E402", "B905", "N818", "S101"]
show-fixes = true
target-version = "py39"
